GTK Reorganize, add custom cursor support (#158)
* first file split

* custom cursor

* fix mock

* lint

* copypasta error
