+++ src/gscreenshot/frontend/gtk/__init__.py
@@ -2,7 +2,7 @@
 #pylint: disable=wrong-import-order
 #pylint: disable=wrong-import-position
 #pylint: disable=ungrouped-imports
-#pylint: disable=too-many-lines
+#pylint: disable=too-many-statements
 '''
 Classes for the GTK gscreenshot frontend
 '''
@@ -15,7 +15,9 @@ from time import sleep
 from pkg_resources import resource_string, resource_filename
 import pygtkcompat
 from gscreenshot import Gscreenshot
-from gscreenshot.util import GSCapabilities
+from gscreenshot.frontend.gtk.dialogs import OpenWithDialog, WarningDialog
+from gscreenshot.frontend.gtk.dialogs import FileSaveDialog, FileOpenDialog
+from gscreenshot.frontend.gtk.view import View
 from gscreenshot.screenshooter.exceptions import NoSupportedScreenshooterError
 from gscreenshot.screenshot.effects.crop import CropEffect
 
@@ -27,405 +29,6 @@ from gi.repository import GObject
 from gi.repository import GLib
 
 i18n = gettext.gettext
-
-
-class View(object):
-    '''View class for the GTK frontend'''
-
-    def __init__(self, window, builder, capabilities):
-        self._builder = builder
-        self._window = window
-        self._window_is_fullscreen:bool = False
-        self._was_maximized:bool = False
-
-        self.hide()
-
-        self._capabilities = capabilities
-        self._last_window_dimensions = self._window.get_size()
-        self._header_bar = builder.get_object('header_bar')
-        self._preview:Gtk.Image = builder.get_object('image1')
-        self._control_grid:Gtk.Box = builder.get_object('control_box')
-        self._cursor_selection_items:Gtk.ListStore = \
-            builder.get_object('cursor_selection_items')
-        self._cursor_selection_dropdown:Gtk.ComboBox = \
-            builder.get_object('pointer_selection_dropdown')
-        self._cursor_selection_label:Gtk.Label = \
-            builder.get_object('pointer_selection_label')
-        self._actions_menu:Gtk.Menu = builder.get_object('menu_saveas_additional_actions')
-        self._status_icon:Gtk.Image = builder.get_object('status_icon')
-        self._preview_overlay:Gtk.Overlay = builder.get_object('image_overlay')
-
-        self._prev_btn:Gtk.Button = Gtk.Button.new_from_stock(Gtk.STOCK_GO_BACK)
-        self._prev_btn.set_size_request(20, 20)
-        self._prev_btn.set_opacity(0)
-
-        self._next_btn:Gtk.Button = Gtk.Button.new_from_stock(Gtk.STOCK_GO_FORWARD)
-        self._next_btn.set_size_request(20, 20)
-        self._next_btn.set_opacity(0)
-
-        self._preview_control:Gtk.ButtonBox = Gtk.ButtonBox()
-        self._preview_control.set_hexpand(False)
-        self._preview_control.set_vexpand(False)
-        self._preview_control.set_halign(Gtk.Align.CENTER)
-        self._preview_control.set_valign(Gtk.Align.END)
-        self._preview_control.add(self._prev_btn)
-        self._preview_control.add(self._next_btn)
-
-        self._preview_overlay.add_overlay(self._preview_control)
-
-        self.update_gallery_controls(False, False)
-
-        if GSCapabilities.ALTERNATE_CURSOR in self._capabilities:
-            self._init_cursor_combobox()
-
-        if GSCapabilities.WINDOW_SELECTION not in self._capabilities:
-            window_select_button = builder.get_object('button_window')
-            self._disable_and_hide(window_select_button)
-
-        if GSCapabilities.REGION_SELECTION not in self._capabilities:
-            region_select_button = builder.get_object('button_selectarea')
-            self._disable_and_hide(region_select_button)
-
-        if GSCapabilities.CURSOR_CAPTURE not in self._capabilities:
-            checkbox_capture_cursor = builder.get_object('checkbox_capture_cursor')
-            self._disable_and_hide(checkbox_capture_cursor)
-
-        if GSCapabilities.REUSE_REGION not in self._capabilities:
-            reuse_region_dropdown = builder.get_object('selection_actions_btn')
-            selectarea_gtkbox = builder.get_object('select_area_gtkbox')
-            self._disable_and_hide(reuse_region_dropdown)
-            selectarea_gtkbox.remove(reuse_region_dropdown)
-
-    def _disable_and_hide(self, widget):
-        '''disables and hides a widget'''
-        widget.set_opacity(0)
-        widget.set_sensitive(0)
-
-    def _enable_and_show(self, widget):
-        '''enables and shows a widget'''
-        widget.set_opacity(1)
-        widget.set_sensitive(1)
-
-    def _hover_effect(self, widget, *_):
-        '''
-        applies a higher opacity to the widget when the cursor hovers
-        '''
-        widget.set_opacity(.9)
-
-    def _unhover_effect(self, widget, *_):
-        '''
-        applies a lighter opacity to the widget when the cursor leaves
-        '''
-        widget.set_opacity(.4)
-
-    def update_gallery_controls(self, show_next=True, show_previous=True):
-        '''
-        updates the preview controls to match the current state
-        '''
-        while Gtk.events_pending():
-            Gtk.main_iteration()
-
-        if show_next and self._next_btn.get_opacity() <= 0:
-            self._next_btn.set_opacity(.5)
-            self._next_btn.connect('enter', self._hover_effect)
-            self._next_btn.connect('leave', self._unhover_effect)
-        elif not show_next and self._next_btn.get_opacity() > 0:
-            self._next_btn.set_opacity(0)
-            self._next_btn.disconnect_by_func(self._hover_effect)
-            self._next_btn.disconnect_by_func(self._unhover_effect)
-
-        if show_previous and self._prev_btn.get_opacity() <= 0:
-            self._prev_btn.set_opacity(.5)
-            self._prev_btn.connect('enter', self._hover_effect)
-            self._prev_btn.connect('leave', self._unhover_effect)
-        elif not show_previous and self._prev_btn.get_opacity() > 0:
-            self._prev_btn.set_opacity(0)
-            self._prev_btn.disconnect_by_func(self._hover_effect)
-            self._prev_btn.disconnect_by_func(self._unhover_effect)
-
-    def connect_signals(self, presenter):
-        '''
-        connects signals
-        '''
-        self._builder.connect_signals(presenter)
-
-        self._window.connect("check-resize", presenter.on_window_resize)
-        self._window.connect("window-state-event", presenter.window_state_event_handler)
-        self._window.set_icon_from_file(
-            resource_filename('gscreenshot.resources.pixmaps', 'gscreenshot.png')
-        )
-
-        if self._prev_btn is not None:
-            self._prev_btn.connect('clicked', presenter.on_preview_prev_clicked)
-
-        if self._next_btn is not None:
-            self._next_btn.connect('clicked', presenter.on_preview_next_clicked)
-
-    def flash_status_icon(self, stock_name: str="emblem-ok"):
-        """
-        Shows the status/confirmation icon in the UI after an action
-        for a second.
-
-        By default this will show the "ok" icon but another one can be
-        passed also.
-
-        This is non-blocking.
-        """
-
-        self._status_icon.set_from_icon_name(stock_name, Gtk.IconSize.BUTTON)
-        self._enable_and_show(self._status_icon)
-
-        while Gtk.events_pending():
-            Gtk.main_iteration()
-
-        _thread = threading.Thread(target=self._finish_flash_status_icon)
-        _thread.daemon = True
-        _thread.start()
-
-    def _finish_flash_status_icon(self):
-        """
-        Lets the status icon be visible for a second, then hides the widget
-        by setting its opacity to 0.
-        """
-        sleep(1)
-        self._disable_and_hide(self._status_icon)
-
-    def set_busy(self):
-        """
-        Sets the window as busy with visual indicators
-        """
-        cursor = Gdk.Cursor.new(Gdk.CursorType.WATCH)
-        self._window.get_window().set_cursor(cursor)
-        while Gtk.events_pending():
-            Gtk.main_iteration()
-
-    def set_ready(self):
-        """
-        Sets the window as ready with visual indicators
-        """
-        self._window.get_window().set_cursor(None)
-        while Gtk.events_pending():
-            Gtk.main_iteration()
-
-    def _init_cursor_combobox(self):
-        combo = self._cursor_selection_dropdown
-        renderer = Gtk.CellRendererPixbuf()
-        combo.pack_start(renderer, False)
-        combo.add_attribute(renderer, "pixbuf", 0)
-
-        renderer = Gtk.CellRendererText()
-        combo.pack_start(renderer, True)
-        combo.add_attribute(renderer, "text", 1)
-
-    def show_cursor_options(self, show: bool):
-        '''
-        Toggle the cursor combobox and label hidden/visible
-        '''
-        if show and GSCapabilities.ALTERNATE_CURSOR in self._capabilities:
-            self._enable_and_show(self._cursor_selection_dropdown)
-            self._enable_and_show(self._cursor_selection_label)
-        else:
-            self._disable_and_hide(self._cursor_selection_dropdown)
-            self._disable_and_hide(self._cursor_selection_label)
-
-    def update_available_cursors(self, cursors: dict):
-        '''
-        Update the available cursor selection in the combolist
-        Params: self, {name: PIL.Image}
-        '''
-        self._cursor_selection_items.clear()
-        for cursor_name in cursors:
-            if cursors[cursor_name] is not None:
-                descriptor = io.BytesIO()
-                image = cursors[cursor_name]
-                image.thumbnail((
-                    self._cursor_selection_dropdown.get_allocation().height*.42,
-                    self._cursor_selection_dropdown.get_allocation().width*.42
-                ))
-                image.save(descriptor, "png")
-                contents = descriptor.getvalue()
-                descriptor.close()
-                loader = Gtk.gdk.PixbufLoader("png")
-                loader.write(contents)
-                pixbuf = loader.get_pixbuf()
-                loader.close()
-
-                self._cursor_selection_items.append(
-                    [pixbuf, i18n('cursor-' + cursor_name), cursor_name]
-                )
-            else:
-                self._cursor_selection_items.append(
-                    [None, i18n('cursor-' + cursor_name), cursor_name]
-                )
-
-            if cursor_name == "theme":
-                self._cursor_selection_dropdown.set_active(
-                    len(self._cursor_selection_items)-1
-                )
-
-    def run(self):
-        '''Run the view'''
-        self._window.set_position(Gtk.WIN_POS_CENTER)
-        # Set the initial size of the window
-        active_window = Gdk.get_default_root_window()
-        while active_window is None:
-            # There appears to be a race condition with getting the active window,
-            # so we'll keep trying until we have it
-            active_window = Gdk.get_default_root_window()
-
-        initial_screen = self._window.get_screen().get_monitor_at_window(active_window)
-        geometry = self._window.get_screen().get_monitor_geometry(initial_screen)
-
-        if self._header_bar is not None:
-            height_x = .6
-        else:
-            height_x = .48
-
-        gscreenshot_height = geometry.height * height_x
-        gscreenshot_width = gscreenshot_height * .9
-
-        if geometry.height > geometry.width:
-            gscreenshot_width = geometry.width * height_x
-            gscreenshot_height = gscreenshot_width * .9
-
-        self._window.set_size_request(gscreenshot_width, gscreenshot_height)
-
-        self._window.show_all()
-
-    def show_actions_menu(self):
-        '''
-        Show the actions/saveas menu at the pointer
-        '''
-        self._actions_menu.popup_at_pointer()
-
-    def get_window(self):
-        '''Returns the associated window'''
-        return self._window
-
-    def toggle_fullscreen(self):
-        '''Toggle the window to full screen'''
-        if self._window_is_fullscreen:
-            self._window.unfullscreen()
-        else:
-            self._window.fullscreen()
-
-        self._window_is_fullscreen = not self._window_is_fullscreen
-
-    def handle_state_event(self, widget, event):
-        '''Handles a window state event'''
-        widget = None
-        self._was_maximized = bool(event.new_window_state & Gtk.gdk.WINDOW_STATE_MAXIMIZED)
-        self._window_is_fullscreen = bool(
-                            Gtk.gdk.WINDOW_STATE_FULLSCREEN & event.new_window_state)
-
-    def hide(self):
-        '''Hide the view'''
-        self._window.set_geometry_hints(None, min_width=-1, min_height=-1)
-        # We set the opacity to 0 because hiding the window is
-        # subject to window closing effects, which can take long
-        # enough that the window will still appear in the screenshot
-        self._window.set_opacity(0)
-
-        # This extra step allows the window to be unmaximized after it
-        # reappears. Otherwise, the hide() call clears the previous
-        # state and the window is stuck maximized. We restore the
-        # maximization when we unhide the window.
-        if self._was_maximized:
-            self._window.unmaximize()
-
-        self._window.hide()
-
-        while Gtk.events_pending():
-            Gtk.main_iteration()
-
-        sleep(0.2)
-
-    def unhide(self):
-        '''Unhide the view'''
-        self._window.set_sensitive(True)
-        self._window.set_opacity(1)
-
-        original_window_size = self._window.get_size()
-        self._window.set_geometry_hints(
-            None,
-            min_width=original_window_size.width,
-            min_height=original_window_size.height
-        )
-
-        if self._was_maximized:
-            self._window.maximize()
-
-        self._window.show_all()
-
-    def resize(self):
-        '''Resize the display'''
-        current_window_size = self._window.get_size()
-        if self._last_window_dimensions is None:
-            self._last_window_dimensions = current_window_size
-
-        if (self._last_window_dimensions.width != current_window_size.width
-                or self._last_window_dimensions.height != current_window_size.height):
-
-            self._last_window_dimensions = current_window_size
-
-    def get_preview_dimensions(self):
-        '''Get the current dimensions of the preview widget'''
-        window_size = self._window.get_size()
-        control_size = self._control_grid.get_allocation()
-
-        header_height = 0
-        if self._header_bar is not None:
-            header_height = self._header_bar.get_allocation().height
-
-        width_x = .8 if self._header_bar is not None else .98
-
-        preview_size = (
-            (window_size.height-control_size.height-(.6*header_height))*.98,
-            window_size.width*width_x
-        )
-
-        height = preview_size[0]
-        width = preview_size[1]
-
-        return height, width
-
-    def update_preview(self, pixbuf):
-        '''
-        Update the preview widget with a new image.
-
-        This assumes the pixbuf has already been resized appropriately.
-        '''
-        # view the previewPixbuf in the image_preview widget
-        self._preview.set_from_pixbuf(pixbuf)
-
-    def copy_to_clipboard(self, pixbuf):
-        """
-        Copy the provided image to the screen's clipboard,
-        if it supports persistence
-        """
-        clipboard = Gtk.Clipboard.get(Gdk.SELECTION_CLIPBOARD)
-        display = Gdk.Display.get_default()
-
-        if display.supports_clipboard_persistence():
-            clipboard.set_image(pixbuf)
-            clipboard.store()
-            return True
-
-        return False
-
-    def run_dialog(self, dialog):
-        '''Run a dialog window and return the outcome'''
-        self._window.set_sensitive(False)
-        result = dialog.run()
-        self._window.set_sensitive(True)
-
-        try:
-            dialog.destroy()
-        except AttributeError:
-            # This process is wonky, and due to incorrect polymorphism
-            pass
-
-        return result
 
 
 class Presenter(object):
@@ -456,6 +59,8 @@ class Presenter(object):
         self._overwrite_mode = True
 
         cursors = self._app.get_available_cursors()
+        cursors[i18n("custom")] = None
+
         self._cursor_selection = 'theme'
 
         self._view.update_available_cursors(
@@ -540,7 +145,50 @@ class Presenter(object):
 
     def selected_cursor_changed(self, widget):
         '''Handle a change to the selected cursor'''
-        self._cursor_selection = widget.get_model()[widget.get_active()][2]
+        try:
+            cursor_selection = widget.get_model()[widget.get_active()][2]
+        except IndexError:
+            return
+
+        if cursor_selection == "custom":
+
+            file_filter:Gtk.FileFilter = Gtk.FileFilter()
+            supported_formats = self._app.get_supported_formats()
+            _ = [file_filter.add_mime_type(
+                f"image/{format}") for format in supported_formats
+                if format not in ["pdf"]
+            ]
+
+            chooser = FileOpenDialog(
+                file_filter=file_filter
+            )
+            chosen = None
+            cancelled = False
+            while not (chosen or cancelled):
+                chosen = self._view.run_dialog(chooser)
+                if chosen is None:
+                    cancelled = True
+
+            if chosen:
+                try:
+                    cursor_name = self._app.register_stamp_image(chosen)
+                #pylint: disable=broad-except
+                except Exception:
+                    warning = WarningDialog(f"Unable to open {chosen}")
+                    self._view.run_dialog(warning)
+                    return
+
+                cursors = self._app.get_available_cursors()
+                cursors[i18n("custom")] = None
+
+                self._view.update_available_cursors(
+                    cursors,
+                    cursor_name
+                )
+                if cursor_name:
+                    cursor_selection = cursor_name
+
+        self._cursor_selection = cursor_selection
 
     def on_button_all_clicked(self, *_):
         '''Take a screenshot of the full screen'''
@@ -600,6 +248,21 @@ class Presenter(object):
             show_next=screenshot_collection.has_next(),
             show_previous=screenshot_collection.has_previous()
         )
+
+    def effect_checkbox_handler(self, widget, effect):
+        '''
+        Handles toggling effects on and off
+        '''
+        screenshot = self._app.get_screenshot_collection().cursor_current()
+        if screenshot is None:
+            return
+
+        if widget.get_active():
+            effect.enable()
+        else:
+            effect.disable()
+
+        self._show_preview()
 
     def on_button_saveas_clicked(self, *_):
         '''Handle the saveas button'''
@@ -837,101 +500,6 @@ class Presenter(object):
         preview_img = self._app.get_thumbnail(width, height, with_border=True)
 
         self._view.update_preview(self._image_to_pixbuf(preview_img))
-
-
-class OpenWithDialog(Gtk.AppChooserDialog):
-    '''The "Open With" dialog'''
-
-    def __init__(self, parent=None):
-
-        Gtk.AppChooserDialog.__init__(self, content_type="image/png", parent=parent)
-        self.set_title(i18n("Choose an Application"))
-        self.connect("response", self._on_response)
-        self.appinfo = None
-
-    def _on_response(self, dialog, response):
-        if response == Gtk.ResponseType.OK:
-            self.appinfo = self.get_app_info()
-        else:
-            self.appinfo = None
-
-
-class FileSaveDialog(object):
-    '''The 'save as' dialog'''
-    def __init__(self, default_filename=None, default_folder=None,
-        parent=None, choose_directory=False
-    ):
-        self.default_filename = default_filename
-        self.default_folder = default_folder
-        self.parent = parent
-        self._choose_directory = choose_directory
-
-    def run(self):
-        ''' Run the dialog'''
-        filename = self.request_file()
-
-        return filename
-
-    def request_file(self):
-        '''Run the file selection dialog'''
-        action = Gtk.FILE_CHOOSER_ACTION_SAVE
-        if self._choose_directory:
-            action = Gtk.FILE_CHOOSER_ACTION_CREATE_FOLDER
-
-        chooser = Gtk.FileChooserDialog(
-                transient_for=self.parent,
-                title=None,
-                action=action,
-                buttons=(
-                    Gtk.STOCK_CANCEL,
-                    Gtk.RESPONSE_CANCEL,
-                    Gtk.STOCK_SAVE,
-                    Gtk.RESPONSE_OK
-                    )
-                )
-
-        if self.default_filename is not None:
-            chooser.set_current_name(self.default_filename)
-
-        if self.default_folder is not None:
-            chooser.set_current_folder(self.default_folder)
-
-        chooser.set_do_overwrite_confirmation(True)
-
-        response = chooser.run()
-
-        if response == Gtk.RESPONSE_OK:
-            return_value = chooser.get_filename()
-        else:
-            return_value = None
-
-        chooser.destroy()
-        return return_value
-
-
-class WarningDialog():
-    '''A warning dialog'''
-
-    def __init__(self, message, parent=None):
-        self.parent = parent
-        self.message_dialog = Gtk.MessageDialog(
-                parent,
-                None,
-                Gtk.MESSAGE_WARNING,
-                Gtk.BUTTONS_OK,
-                message
-                )
-
-    def run(self):
-        '''Run the warning dialog'''
-        if self.parent is not None:
-            self.parent.set_sensitive(False)
-
-        self.message_dialog.run()
-        self.message_dialog.destroy()
-
-        if self.parent is not None:
-            self.parent.set_sensitive(True)
 
 
 def main():
